https://github.com/AdguardTeam/AdguardBrowserExtension/issues/8
