+++ Extension/browser/chrome/chrome_background.html
@@ -3,12 +3,10 @@
 <head>
 
     <script type="text/javascript" src="lib/libs/deferred.min.js"></script>
-
     <script type="text/javascript" src="lib/content-script/common-script.js"></script>
     <script type="text/javascript" src="lib/background-page.js"></script>
     <script type="text/javascript" src="require.js"></script>
     <script type="text/javascript" src="prefs.js"></script>
-
     <script type="text/javascript" src="lib/utils/sha256.patched.js"></script>
     <script type="text/javascript" src="lib/utils/promises.js"></script>
     <script type="text/javascript" src="lib/utils/punycode.js"></script>
@@ -26,27 +24,17 @@
     <script type="text/javascript" src="lib/utils/notifier.js"></script>
     <script type="text/javascript" src="lib/utils/frames.js"></script>
     <script type="text/javascript" src="lib/utils/page-stats.js"></script>
-
     <script type="text/javascript" src="lib/filter/rules/shortcuts-lookup-table.js"></script>
-
     <script type="text/javascript" src="lib/filter/rules/base-filter-rule.js"></script>
-    <script type="text/javascript" src="lib/filter/rules/filter-classes.js"></script>
-
     <script type="text/javascript" src="lib/filter/rules/css-filter-rule.js"></script>
     <script type="text/javascript" src="lib/filter/rules/css-filter.js"></script>
-
     <script type="text/javascript" src="lib/filter/rules/script-filter-rule.js"></script>
     <script type="text/javascript" src="lib/filter/rules/script-filter.js"></script>
-
     <script type="text/javascript" src="lib/filter/rules/url-filter-rule.js"></script>
     <script type="text/javascript" src="lib/filter/rules/url-filter.js"></script>
-
     <script type="text/javascript" src="lib/filter/safebrowsing-filter.js"></script>
-
     <script type="text/javascript" src="lib/filter/whitelist.js"></script>
-
     <script type="text/javascript" src="lib/filter/filters.js"></script>
-
     <script type="text/javascript" src="lib/filter/filters-hit.js"></script>
     <script type="text/javascript" src="lib/filter/storage.js"></script>
     <script type="text/javascript" src="lib/filter/locale-detect.js"></script>
@@ -56,9 +44,7 @@
     <script type="text/javascript" src="lib/filter/integration.js"></script>
     <script type="text/javascript" src="lib/filter/filtering-log.js"></script>
     <script type="text/javascript" src="lib/filter/request-blocking.js"></script>
-
     <script type="text/javascript" src="lib/content-message-handler.js"></script>
-
     <script type="text/javascript" src="ui.js"></script>
     <script type="text/javascript" src="background.js"></script>
     <script type="text/javascript" src="webrequest.js"></script>
