+++ Extension/browser/firefox_legacy/lib/customizableUI.js
@@ -21,8 +21,8 @@
  */
 
 
-var ConcurrentUtils = require('../lib/utils/browser-utils').ConcurrentUtils;
-var {UiUtils, WindowObserver} = require('../lib/uiUtils');
+var ConcurrentUtils = require('./utils/browser-utils').ConcurrentUtils;
+var {UiUtils, WindowObserver} = require('./uiUtils');
 
 var widgets = Object.create(null);
