+++ Extension/lib/filter/integration.js
@@ -19,13 +19,12 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var userSettings = require('../lib/utils/user-settings').userSettings;
-var ServiceClient = require('../lib/utils/service-client').ServiceClient;
-var FilterRule = require('../lib/filter/rules/base-filter-rule').FilterRule;
-require('../lib/filter/rules/filter-classes');
-var UrlFilterRule = require('../lib/filter/rules/url-filter-rule').UrlFilterRule;
-var StringUtils = require('../lib/utils/common').StringUtils;
-var RequestTypes = require('../lib/utils/common').RequestTypes;
+var userSettings = require('../../lib/utils/user-settings').userSettings;
+var ServiceClient = require('../../lib/utils/service-client').ServiceClient;
+var FilterRule = require('../../lib/filter/rules/base-filter-rule').FilterRule;
+var UrlFilterRule = require('../../lib/filter/rules/url-filter-rule').UrlFilterRule;
+var StringUtils = require('../../lib/utils/common').StringUtils;
+var RequestTypes = require('../../lib/utils/common').RequestTypes;
 
 /**
  * AdguardApplication is used for integration of Adguard extension and Adguard for Windows/Mac/Android versions.
