+++ Extension/lib/filter/locale-detect.js
@@ -19,12 +19,12 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var Utils = require('../lib/utils/browser-utils').Utils;
-var Prefs = require('./prefs').Prefs;
-var UrlUtils = require('../lib/utils/url').UrlUtils;
-var userSettings = require('../lib/utils/user-settings').userSettings;
-var EventNotifier = require('../lib/utils/notifier').EventNotifier;
-var EventNotifierTypes = require('../lib/utils/common').EventNotifierTypes;
+var Utils = require('../../lib/utils/browser-utils').Utils;
+var Prefs = require('../../lib/prefs').Prefs;
+var UrlUtils = require('../../lib/utils/url').UrlUtils;
+var userSettings = require('../../lib/utils/user-settings').userSettings;
+var EventNotifier = require('../../lib/utils/notifier').EventNotifier;
+var EventNotifierTypes = require('../../lib/utils/common').EventNotifierTypes;
 
 /**
  * Initialize LocaleDetectService.
