+++ Extension/lib/filter/rules/url-filter.js
@@ -19,10 +19,10 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var ShortcutsLookupTable = require('../lib/filter/rules/shortcuts-lookup-table').ShortcutsLookupTable;
-var StringUtils = require('../lib/utils/common').StringUtils;
-var CollectionUtils = require('../lib/utils/common').CollectionUtils;
-var UrlUtils = require('../lib/utils/url').UrlUtils;
+var ShortcutsLookupTable = require('../../../lib/filter/rules/shortcuts-lookup-table').ShortcutsLookupTable;
+var StringUtils = require('../../../lib/utils/common').StringUtils;
+var CollectionUtils = require('../../../lib/utils/common').CollectionUtils;
+var UrlUtils = require('../../../lib/utils/url').UrlUtils;
 
 /**
  * Filter for Url filter rules.
