+++ Extension/lib/filter/safebrowsing-filter.js
@@ -19,14 +19,14 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var ServiceClient = require('../lib/utils/service-client').ServiceClient;
-var LocalStorageCache = require('../lib/utils/cache').LocalStorageCache;
-var UrlUtils = require('../lib/utils/url').UrlUtils;
-var Log = require('../lib/utils/log').Log;
-var LS = require('../lib/utils/local-storage').LS;
-var Prefs = require('../lib/prefs').Prefs;
-var StringUtils = require('../lib/utils/common').StringUtils;
-var SHA256 = require('../lib/utils/sha256.patched').SHA256;
+var ServiceClient = require('../../lib/utils/service-client').ServiceClient;
+var LocalStorageCache = require('../../lib/utils/cache').LocalStorageCache;
+var UrlUtils = require('../../lib/utils/url').UrlUtils;
+var Log = require('../../lib/utils/log').Log;
+var LS = require('../../lib/utils/local-storage').LS;
+var Prefs = require('../../lib/prefs').Prefs;
+var StringUtils = require('../../lib/utils/common').StringUtils;
+var SHA256 = require('../../lib/utils/sha256.patched').SHA256;
 
 /**
  * Initializing SafebrowsingFilter.
