+++ Extension/lib/filter/update-service.js
@@ -19,16 +19,16 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var Log = require('../lib/utils/log').Log;
-var Utils = require('../lib/utils/browser-utils').Utils;
-var LS = require('../lib/utils/local-storage').LS;
-var Prefs = require('../lib/prefs').Prefs;
-var AntiBannerFiltersId = require('../lib/utils/common').AntiBannerFiltersId;
-var FS = require('../lib/utils/file-storage').FS;
-var FilterStorage = require('../lib/filter/storage').FilterStorage;
-var CollectionUtils = require('../lib/utils/common').CollectionUtils;
-var Promise = require('../lib/utils/promises').Promise;
-var filterRulesHitCount = require('../lib/filter/filters-hit').filterRulesHitCount;
+var Log = require('../../lib/utils/log').Log;
+var Utils = require('../../lib/utils/browser-utils').Utils;
+var LS = require('../../lib/utils/local-storage').LS;
+var Prefs = require('../../lib/prefs').Prefs;
+var AntiBannerFiltersId = require('../../lib/utils/common').AntiBannerFiltersId;
+var FS = require('../../lib/utils/file-storage').FS;
+var FilterStorage = require('../../lib/filter/storage').FilterStorage;
+var CollectionUtils = require('../../lib/utils/common').CollectionUtils;
+var Promise = require('../../lib/utils/promises').Promise;
+var filterRulesHitCount = require('../../lib/filter/filters-hit').filterRulesHitCount;
 var simpleStorage = require('sdk/simple-storage');
 
 /**
