+++ Extension/lib/filter/whitelist.js
@@ -15,18 +15,17 @@
  * along with Adguard Browser Extension.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-var Utils = require('../lib/utils/browser-utils').Utils;
-var StringUtils = require('../lib/utils/common').StringUtils;
-var UrlUtils = require('../lib/utils/url').UrlUtils;
-var CollectionUtils = require('../lib/utils/common').CollectionUtils;
-var AntiBannerFiltersId = require('../lib/utils/common').AntiBannerFiltersId;
-var RequestTypes = require('../lib/utils/common').RequestTypes;
-var LS = require('../lib/utils/local-storage').LS;
-var userSettings = require('../lib/utils/user-settings').userSettings;
+var Utils = require('../../lib/utils/browser-utils').Utils;
+var StringUtils = require('../../lib/utils/common').StringUtils;
+var UrlUtils = require('../../lib/utils/url').UrlUtils;
+var CollectionUtils = require('../../lib/utils/common').CollectionUtils;
+var AntiBannerFiltersId = require('../../lib/utils/common').AntiBannerFiltersId;
+var RequestTypes = require('../../lib/utils/common').RequestTypes;
+var LS = require('../../lib/utils/local-storage').LS;
+var userSettings = require('../../lib/utils/user-settings').userSettings;
 
-var FilterRule = require('../lib/filter/rules/base-filter-rule').FilterRule;
-require('../lib/filter/rules/filter-classes');
-var UrlFilter = require('../lib/filter/rules/url-filter').UrlFilter;
+var FilterRule = require('../../lib/filter/rules/base-filter-rule').FilterRule;
+var UrlFilter = require('../../lib/filter/rules/url-filter').UrlFilter;
 
 var allowAllWhiteListRule = {
     whiteListRule: true,
