+++ Extension/lib/utils/browser-utils.js
@@ -1,13 +1,30 @@
+/**
+ * This file is part of Adguard Browser Extension (https://github.com/AdguardTeam/AdguardBrowserExtension).
+ *
+ * Adguard Browser Extension is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Adguard Browser Extension is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Adguard Browser Extension.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 var Cu = require('chrome').Cu;
 var Cc = require('chrome').Cc;
 var Ci = require('chrome').Ci;
-var LS = require('../lib/utils/local-storage').LS;
-var Prefs = require('../lib/prefs').Prefs;
 var setTimeout = require('sdk/timers').setTimeout;
 var clearTimeout = require('sdk/timers').clearTimeout;
 Cu.import("resource://gre/modules/Services.jsm");
 
-var Log = require('../lib/utils/log').Log;
+var LS = require('../../lib/utils/local-storage').LS;
+var Prefs = require('../../lib/prefs').Prefs;
+var Log = require('../../lib/utils/log').Log;
 
 var Utils = exports.Utils = {
