+++ Extension/lib/utils/service-client.js
@@ -19,14 +19,14 @@
  * Initializing required libraries for this file.
  * require method is overridden in Chrome extension (port/require.js).
  */
-var Log = require('../lib/utils/log').Log;
-var Utils = require('../lib/utils/browser-utils').Utils;
-var FilterRule = require('../lib/filter/rules/base-filter-rule').FilterRule;
-require('../lib/filter/rules/filter-classes');
-var Prefs = require('../lib/prefs').Prefs;
 var Cc = require('chrome').Cc;
 var Ci = require('chrome').Ci;
 var XMLHttpRequestConstructor = Cc["@mozilla.org/xmlextras/xmlhttprequest;1"];
+
+var Log = require('../../lib/utils/log').Log;
+var Utils = require('../../lib/utils/browser-utils').Utils;
+var FilterRule = require('../../lib/filter/rules/base-filter-rule').FilterRule;
+var Prefs = require('../../lib/prefs').Prefs;
 
 /**
  * Class for working with our backend server.
@@ -108,7 +108,7 @@ ServiceClient.prototype = {
 			return;
 		}
 
-		var AdguardFilterVersion = require('../lib/filter/antibanner').AdguardFilterVersion;
+		var AdguardFilterVersion = require('../../lib/filter/antibanner').AdguardFilterVersion;
 
 		var success = function (response) {
 			var xml = response.responseXML;
@@ -147,7 +147,7 @@ ServiceClient.prototype = {
 	 */
 	loadFilterRules: function (filterId, useOptimizedFilters, successCallback, errorCallback) {
 
-		var AdguardFilterVersion = require('../lib/filter/antibanner').AdguardFilterVersion;
+		var AdguardFilterVersion = require('../../lib/filter/antibanner').AdguardFilterVersion;
 
 		var success = function (response) {
