+++ Extension/tests/test.html
@@ -22,7 +22,6 @@
     <script type="text/javascript" src="../lib/filter/rules/script-filter.js"></script>
     <script type="text/javascript" src="../lib/filter/rules/url-filter-rule.js"></script>
     <script type="text/javascript" src="../lib/filter/rules/url-filter.js"></script>
-    <script type="text/javascript" src="../lib/filter/rules/filter-classes.js"></script>
 
     <script type="text/javascript" src="../browser/chrome/lib/utils/local-storage.js"></script>
     <script type="text/javascript" src="../lib/utils/cache.js"></script>
